[Aniket] | BAH-3192 | Add. wa.templates package, Moved templates method to this package, Removed files not in use
