--- src/main/java/org/bahmni/whatsapp/appointments/contract/patient/OpenMRSPatient.java
@@ -1,40 +1,0 @@
-package org.bahmni.whatsapp.appointments.contract.patient;
-public class OpenMRSPatient {
-    private String patientId;
-    private String givenName;
-    private String middleName;
-    private String familyName;
-
-    public String getFamilyName() {
-        return familyName;
-    }
-
-    public void setFamilyName(String familyName) {
-        this.familyName = familyName;
-    }
-
-    public String getGivenName() {
-        return givenName;
-    }
-
-    public void setGivenName(String givenName) {
-        this.givenName = givenName;
-    }
-
-    public String getMiddleName() {
-        return middleName;
-    }
-
-    public void setMiddleName(String middleName) {
-        this.middleName = middleName;
-    }
-
-    public String getPatientId() {
-        return patientId;
-    }
-
-    public void setPatientId(String patientId) {
-        this.patientId = patientId;
-    }
-
-}
