--- src/main/java/org/bahmni/whatsapp/appointments/contract/patient/OpenMRSPersonAttributeType.java
@@ -1,33 +1,0 @@
-package org.bahmni.whatsapp.appointments.contract.patient;
-
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-
-@JsonIgnoreProperties(ignoreUnknown = true)
-public class OpenMRSPersonAttributeType {
-    private String uuid;
-    private String display;
-
-    public String getUuid() {
-        return uuid;
-    }
-
-    public void setUuid(String uuid) {
-        this.uuid = uuid;
-    }
-
-    public String getDisplay() {
-        return display;
-    }
-
-    public void setDisplay(String display) {
-        this.display = display;
-    }
-
-    @Override
-    public String toString() {
-        return "OpenMRSPersonAttributeType{" +
-                "uuid='" + uuid + '\'' +
-                ", display='" + display + '\'' +
-                '}';
-    }
-}
