--- src/main/java/org/bahmni/whatsapp/appointments/contract/patient/OpenMRSPersonName.java
@@ -1,70 +1,0 @@
-package org.bahmni.whatsapp.appointments.contract.patient;
-
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-
-
-@JsonIgnoreProperties(ignoreUnknown = true)
-public class OpenMRSPersonName {
-    private String uuid;
-    private String givenName;
-    private String middleName;
-    private String familyName;
-    private boolean voided;
-
-    public String getUuid() {
-        return uuid;
-    }
-
-    public void setUuid(String uuid) {
-        this.uuid = uuid;
-    }
-
-    public String getGivenName() {
-        return givenName;
-    }
-
-    public void setGivenName(String givenName) {
-        this.givenName = givenName;
-    }
-
-    public String getMiddleName() {
-        return middleName;
-    }
-
-    public void setMiddleName(String middleName) {
-        this.middleName = middleName;
-    }
-
-    public String getFamilyName() {
-        return familyName;
-    }
-
-    public void setFamilyName(String familyName) {
-        this.familyName = familyName;
-    }
-
-    public boolean isVoided() {
-        return voided;
-    }
-
-    public void setVoided(boolean voided) {
-        this.voided = voided;
-    }
-
-    @Override
-    public String toString() {
-        return "OpenMRSPersonName{" +
-                "uuid='" + uuid + '\'' +
-                ", givenName='" + givenName + '\'' +
-                ", middleName='" + middleName + '\'' +
-                ", familyName='" + familyName + '\'' +
-                ", voided=" + voided +
-                '}';
-    }
-
-    public String getPreferredFullName() {
-        return this.getGivenName()
-                + (this.getMiddleName() != null ? " " + this.getMiddleName() : "")
-                + " " + this.getFamilyName();
-    }
-}
