--- src/main/java/org/bahmni/whatsapp/appointments/jobs/FeedJob.java
@@ -1,5 +1,0 @@
-package org.bahmni.whatsapp.appointments.jobs;
-
-public interface FeedJob {
-    void process() throws InterruptedException;
-}
